+++ packages/tsurlfilter/test/rules/network-rule.test.ts
@@ -248,26 +248,16 @@ describe('NetworkRule constructor', () => {
         }).not.toThrow();
 
         expect(() => {
+            new NetworkRule('||*/ads^$to=evil.com', 0);
+        }).not.toThrow();
+
+        expect(() => {
             new NetworkRule('||baddomain.com^$to=', 0);
         }).toThrow(new SyntaxError('$to modifier value cannot be empty'));
 
         expect(() => {
             new NetworkRule('||baddomain.com^$to=example.org|', 0);
         }).toThrow(new SyntaxError('Empty domain specified in "example.org|"'));
-    });
-
-    it('thorws error if $to modfiier value is invalid', () => {
-        expect(() => {
-            new NetworkRule('||*/ads^$to=evil.com', 0);
-        }).not.toThrow();
-
-        expect(() => {
-            new NetworkRule('|*/ads^$to=', 0);
-        }).toThrow(new SyntaxError('$to modifier value cannot be empty'));
-
-        expect(() => {
-            new NetworkRule('|*/ads^$to=evil.com|', 0);
-        }).toThrow(new SyntaxError('Empty domain specified in "evil.com|"'));
     });
 
     it('checks removeparam modifier compatibility', () => {
@@ -362,6 +352,18 @@ describe('NetworkRule constructor', () => {
 
         // TODO: add more specific jsonprune tests during the implementation
         // https://github.com/AdguardTeam/tsurlfilter/issues/72
+    });
+
+    it('checks to modifier compatibility', () => {
+        expect(() => {
+            new NetworkRule('/ads$to=good.org,denyallow=good.com', 0);
+        }).toThrow(new SyntaxError('modifier $to is not compatible with $denyallow modifier'));
+    });
+
+    it('checks denyallow modifier compatibility', () => {
+        expect(() => {
+            new NetworkRule('/ads$to=good.org,denyallow=good.com', 0);
+        }).toThrow(new SyntaxError('modifier $to is not compatible with $denyallow modifier'));
     });
 
     it('works when it handles wide rules with $domain properly', () => {
@@ -1209,8 +1211,8 @@ describe('NetworkRule.match', () => {
         let request: Request;
         rule = new NetworkRule('/ads$to=~evil.*|good.*,script', 0);
 
-        expect(rule.getRestrictedToValues()).toHaveLength(1);
-        expect(rule.getPermittedToValues()).toHaveLength(1);
+        expect(rule.getRestrictedToDomains()).toHaveLength(1);
+        expect(rule.getPermittedToDomains()).toHaveLength(1);
 
         // Correctly matches domain that is specified in permitted domains list
         rule = new NetworkRule('/ads^$to=evil.com', 0);