+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
+## UNRELEASED
+
+### Added
+- Support for `$to` modifier in the MV3 converter.
+
 ## [2.1.6] - 2023-08-04
 
 ### Changed
@@ -24,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.1.4] - 2023-07-13
 
 ### Added
-- Support for $csp modifier in the MV3 converter.
+- Support for `$csp` modifier in the MV3 converter.
 - Scheme for converting network rules into declarative rules.
 
 ### Changed